Implements locations view model
Introduces a locations view model to manage location data and camera position.
The view model fetches location data, sets the initial map location, and updates the camera position based on the currently selected location.

Adds LocationsView to display a map with location markers and a header displaying the name of the current location.
Updates ScappApp to use LocationsView and inject the LocationsViewModel as an environment object.

Adds location struct and makes it Identifiable to be used with MapKit.